backend/requirements.txt
fastapi
uvicorn
motor
bcrypt
python-jose
pydantic[email]
passlib